Refactor the Keccak-f permutation to its own package.
The Keccak-f Permute procedure is moved to its own package so that
there can be multiple generic instantiations of it which can
operate on the same Keccak-f state.

This is necessary for implementing new constructs such as
MonkeyDuplex where two variants of the permutation are used with
different numbers of rounds.

The number of rounds is kept as a generic formal parameter, rather
than an argument to a procedure so that other users of the Permute
procedure (e.g. Sponge and Duplex) do not need to know the number
of rounds.