Merge Keccak Chi and Iota procedures to improve performance.
The Iota round is very simple and only operates on the first lane of
the Keccak state. Merging this operation with the preceding
operation - Chi - slightly improves performance since a memory access
is avoided. In the benchmark the performance of SHA3-224 is improved
by 2-3 MiB/s and the cycles per byte falls from 16.00 to 15.69.