Minor sponge optimization, and update benchmark and results.
A loop in the Sponge now uses a cached value for Rate_Of(Ctx) to
avoid recalculating it several times per loop iteration. This
slightly improves performance in the benchmark by about 1 MiB/s in
the best case.

The benchmark for the Keccak permutation is updated to use the
statically compiled permutation, rather than instantiating the
package at runtime to better demonstrate the optimal performance.

README.md is updated with the latest benchmark results.