Implement CSHAKE and KMAC
CSHAKE and KMAC are new derived modes from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds these two
modes.

The code is proven to be free of runtime errors using GNATprove
(SPARK GPL 2015), but as there are no test vectors (yet) there
aren't any tests to verify the correctness of the algorithms.
However, as there is significant reuse of the existing code
(the XOF construction in particular) which ARE tested this gives
some confidence of correctness of certain aspects of the new
algorithms.