Add the subtype Rate_Bits_Number.
This subtype replaces use of the Positive subtype for variables
storing the rate of a sponge (such as the return type of the
Rate_Of function), and constrains the range of the type to the
length permitted by the specific sponge instantiation. This is
useful in some cases for users which use the rate/capacity
information.