Fix edge-case found by SPARK Discovery 2017.
Many preconditions in libkeccak impose constraints on the length of
a Byte_Array, e.g. Pre => Data'Length <= (Natural'Last / 8)
However, the biggest possible length of a Byte_Array exceeds the
range of the Natural type by 1. This commit reduces the maximum
possible Byte_Array length so that its length always fits in
Natural'Range