Implement versions of Left/Right_Encode for bit-lengths.
These new functions are Left_Encode_Bit_Length and
Right_Encode_Bit_Length. These functions operate the same as
Left_Encode and Right_Encode, except that the input length is
interpreted as a number of bytes, but the output represents the
bit length (i.e. the input number is multipied by 8 in the output).

This commit allows us to remove restrictions on input lengths for
parameters to KMAC and Tuple_Hash, where the lengths of some arrays
were restricted to no greater than Natural'Last / 8. These restrictions
have been removed.