Implement cSHAKE, KMAC, and TupleHash (#1)
* Add String -> Byte_Array conversion function

* Implement CSHAKE and KMAC

CSHAKE and KMAC are new derived modes from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds these two
modes.

The code is proven to be free of runtime errors using GNATprove
(SPARK GPL 2015), but as there are no test vectors (yet) there
aren't any tests to verify the correctness of the algorithms.
However, as there is significant reuse of the existing code
(the XOF construction in particular) which ARE tested this gives
some confidence of correctness of certain aspects of the new
algorithms.

* Add String -> Byte_Array conversion function

* Implement CSHAKE and KMAC

CSHAKE and KMAC are new derived modes from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds these two
modes.

The code is proven to be free of runtime errors using GNATprove
(SPARK GPL 2015), but as there are no test vectors (yet) there
aren't any tests to verify the correctness of the algorithms.
However, as there is significant reuse of the existing code
(the XOF construction in particular) which ARE tested this gives
some confidence of correctness of certain aspects of the new
algorithms.

* Fixes after rebase

* Add documentation to describe usage for precomputation.

* KMAC output length param is moved from the Init to Finish subprogram.

The output length is processed by CSHAKE after the variable-length
user data is processed. Therefore, we only need to know the desired
output length when Finish is called, and don't need to store it
for the full lifetime of the Context object.

* Implement Tuple Hash.

Tuple Hash is a new derived mode from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds this new
(experimental) mode to libkeccak.

* Add support for XOF-like output from Tuple Hash and KMAC.

This functionality is defined in the draft document located at:
http://csrc.nist.gov/publications/drafts/800-185/sp800_185_draft.pdf

* Update the CSHAKE function name strings for Tuple Hash & KMAC.

The function name strings for Tuple Hash and KMAC have been changed
in the latest draft. The draft is available at:
http://csrc.nist.gov/publications/drafts/800-185/sp800_185_draft.pdf

Basically, the strings have been changed from "kmac" to "KMAC" and
from "tuplehash" to "TupleHash".

* Add support for arbitrary bit-length messages in CSHAKE.

* Strengthen postcondition of Padding_Zeroes.

* Add String -> Byte_Array conversion function

* Implement CSHAKE and KMAC

CSHAKE and KMAC are new derived modes from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds these two
modes.

The code is proven to be free of runtime errors using GNATprove
(SPARK GPL 2015), but as there are no test vectors (yet) there
aren't any tests to verify the correctness of the algorithms.
However, as there is significant reuse of the existing code
(the XOF construction in particular) which ARE tested this gives
some confidence of correctness of certain aspects of the new
algorithms.

* Fixes after rebase

* Add documentation to describe usage for precomputation.

* KMAC output length param is moved from the Init to Finish subprogram.

The output length is processed by CSHAKE after the variable-length
user data is processed. Therefore, we only need to know the desired
output length when Finish is called, and don't need to store it
for the full lifetime of the Context object.

* Implement Tuple Hash.

Tuple Hash is a new derived mode from SHA-3, proposed by NIST
to the NIST Hash Forum mailing list. This commit adds this new
(experimental) mode to libkeccak.

* Add support for XOF-like output from Tuple Hash and KMAC.

This functionality is defined in the draft document located at:
http://csrc.nist.gov/publications/drafts/800-185/sp800_185_draft.pdf

* Update the CSHAKE function name strings for Tuple Hash & KMAC.

The function name strings for Tuple Hash and KMAC have been changed
in the latest draft. The draft is available at:
http://csrc.nist.gov/publications/drafts/800-185/sp800_185_draft.pdf

Basically, the strings have been changed from "kmac" to "KMAC" and
from "tuplehash" to "TupleHash".

* Add support for arbitrary bit-length messages in CSHAKE.

* Strengthen postcondition of Padding_Zeroes.

* Implement CSHAKE tests and fix bugs to pass the new tests.

The tests were implemented from the samples provided by NIST at:
http://csrc.nist.gov/groups/ST/toolkit/documents/Examples/cSHAKE_samples.pdf

* Implement KMAC tests and fix bugs to pass new tests.

The KMAC tests were obtained from the samples provided by NIST at:
http://csrc.nist.gov/groups/ST/toolkit/documents/Examples/KMAC_samples.pdf

* Implement TupleHash tests and fix bugs to pass new tests.

* Implement versions of Left/Right_Encode for bit-lengths.

These new functions are Left_Encode_Bit_Length and
Right_Encode_Bit_Length. These functions operate the same as
Left_Encode and Right_Encode, except that the input length is
interpreted as a number of bytes, but the output represents the
bit length (i.e. the input number is multipied by 8 in the output).

This commit allows us to remove restrictions on input lengths for
parameters to KMAC and Tuple_Hash, where the lengths of some arrays
were restricted to no greater than Natural'Last / 8. These restrictions
have been removed.

* Add XOF samples to test suite for KMAC and TupleHash

* Add cSHAKE, KMAC, and TupleHash