Merge Rho and Pi operations for an approx. 20% performance improvement.
The old separate implementations of Rho and Pi are still available for
the time being, but they may not be needed so perhaps they could be
removed.

With this change the performance is improved by about 20% on my machine
(Intel Core i7-2630QM). E.g. SHA3-256 performance is improved
from 151 MiB/s to 181 MiB/s.

KAT tests and unit tests have been run to ensure that the Keccak
permutation algorithm is still correct and yields the same results.