Update benchmark results.
It turns out that the previous benchmarks showing 20% performance
improvement were due to the benchmark program not being correctly
rebuilt after making all the modifications. The results included
in the previous commit were actually missing some parts of the
computation, and are not representative of the true performance.

The makefile has been updated so that GPRbuild is correctly invoked
on the benchmark program, even if nothing needs to be rebuilt.