Reorganize source files into arch-specific directories.
The purpose of this reorganization is to permit future
architecture-dependent instantiations of the Keccak permutations.

Currently, only a generic serial implementation, and x86_64 with SSE2
are supported.

The library is built for a specific architecture by setting the ARCH
and (optionally) SIMD flags on the command line, an example of which
is shown below. These flags default to the current architecture, or
'generic' if the architecture name is unknown. For X86_64, SSE2 is
enabled by default.

Example:
make build ARCH=x86_64 SIMD=SSE2