Add --level=0 to GNATprove settings.
For some reason, without --level=0 in the command line for GNATprove
it is unable to prove all properties with some of the
Generic_Parllel_... packages, even though --prover and --proof are
overriden on the command line.

Furthermore, using --steps instead of --timeout also seems to cause
many checks to go unproved, even when --steps is set to a ridiculously
high number (such as 400 million).