Workaround for a bug in GNATprove during flow analysis.
A bug occurred during flow analysis of keccak-parallel_keccak_1600.ads
where it would incorrectly state that in Permute_All_P4_R12 and similar
instantiations "S" must be declared in the Global aspect, even though
"S" is a parameter, not a global.

The workaround is to avoid private declarations for the Parallel_State
type.