Implement parallel Keccak constructions: ParallelHash and KangarooTwelve. (#3)
* Initial support for parallel algorithms (e.g. KangarooTwelve)

This commit adds (completed untested and unproven) supporting algorithms
for parallel constructions of Keccak. This adds a parallel sponge, XOF,
as well as packages to simulate parallel permutations based on the
existing serial implementations.

An initial (again, untested and unproven) implementation of
KangarooTwelve is also added.

* Fix bug in Pad_Single_Block.

There was a bug where in cases where Max_Bit_Len is a multiple of 8
then the last '1' bit is not appended due to a right-shift of a byte
by 8 positions. This has been fixed.

* First working implementation of KangarooTwelve.

The code has been tested (but not GNATproved) against the KangarooTwelve
test vectors from the Keccak Code Package. Automated KAT tests for these
vectors are not yet implemented however.

* Remove debug messages

* Add KangarooTwelve to the benchmark

* Revert changes to the compilation switches

* Minor optimization to Byte_Lanes

The XOR_Bits_Into_State routine was modified to use loops with fixed
bounds. This yields a performance improvement of about 2%. For example,
SHA3-224 performance increased from ~175 MiB/s to 178 MiB/s.

* Use Bit_Len instead of 'Length

* Add checks to prove absence of runtime errors

* First implementation of Keccak-F[1600]x2 using SSE2 instructions.

KangarooTwelve's performance using this implementation is about
522 MiB/s, compared to about 350 MiB/s when using the serial fallback.

The code is still lacking proof.

* Make the vectorized implementation of Keccak a generic package.

The generic package Keccak.Generic_Parallel_KeccakF implements the Keccak
permutation for any type of SIMD instruction set. Functions and types for
the specific instruction set are provided as generic parameters, allowing
the same code to be instantiated for different instruction sets.

* Add per-arch package to define parallel permutation implementations.

The Keccak.Parallel_Keccak_1600 package will be defined for each
architecture configuration to define instances for 2x 4x and 8x
parallel Keccak-f permutations.

Other packages (e.g. KangarooTwelve) can then reference this package
to reference optimized parallel permutations in a cross-platform way.

* Add Global aspects

* Prove some packages with GNATprove

* Prove more code

* Fix instantiation problems in KangarooTwelve

* Prove Generic_KangarooTwelve

* Prove all instances of Generic_Parallel_Sponge

* Use separate data and function sections.

This permits fine-grained removal of unused functions in programs linked
with the library, reducing overall executable size.

* Tweak GNATprove switches

* Enable SPARK_Mode for SSE2 package

* Disable warnings for postcondition in Ghost code.

The warnings here warn that the postcondition does not check the
result of calling the procedure. But the purpose of these procedures
is that they are Ghost code for aiding in proof.

* Use Ada.Execution_Time, and remove processing loop

* Update benchmark results and tidy up the description

* Reorganize source files into arch-specific directories.

The purpose of this reorganization is to permit future
architecture-dependent instantiations of the Keccak permutations.

Currently, only a generic serial implementation, and x86_64 with SSE2
are supported.

The library is built for a specific architecture by setting the ARCH
and (optionally) SIMD flags on the command line, an example of which
is shown below. These flags default to the current architecture, or
'generic' if the architecture name is unknown. For X86_64, SSE2 is
enabled by default.

Example:
make build ARCH=x86_64 SIMD=SSE2

* Update some documentation

* Add unit tests to check parallel sponge.

This also fixes a Constraint_Error found during testing, as the proofs
are not yet fully functional.

* Output average result for each benchmark

* Selectively run benchmarks based on command-line args

* Add procedure to XOR the same data into multiple parallel states.

The existing XOR_Bits_Into_State procedure for parallel Keccak would
split the buffer into multiple chunks, then each chunk is XOR'ed into
a different state.

For parallel cSHAKE we will need the ability to XOR the *same* data
into all parallel states when processing the customization and
function name strings. Hence the need for this new procedure.

* First (untested, but proven) implementation of ParallelHash

* Prove ParallelHash

* Add --level=0 to GNATprove settings.

For some reason, without --level=0 in the command line for GNATprove
it is unable to prove all properties with some of the
Generic_Parllel_... packages, even though --prover and --proof are
overriden on the command line.

Furthermore, using --steps instead of --timeout also seems to cause
many checks to go unproved, even when --steps is set to a ridiculously
high number (such as 400 million).

* Workaround for a bug in GNATprove during flow analysis.

A bug occurred during flow analysis of keccak-parallel_keccak_1600.ads
where it would incorrectly state that in Permute_All_P4_R12 and similar
instantiations "S" must be declared in the Global aspect, even though
"S" is a parameter, not a global.

The workaround is to avoid private declarations for the Parallel_State
type.

* Fix problems found by GNATprove

* Increase limit on input length for K12 and ParallelHash.

Counting the number of input bytes (rather than the number of input
blocks) avoids the need to multiply the number of blocks by the
block size, which needed restrictions to avoid integer overflow.
Thus, by counting bytes we can remove these restrictions so now
the input length is only limited by Long_Long_Integer'Last, which
is typically 2**63 - 1.

* Fix problems found by GNATprove

* Add missing Inline aspects

* Update unit test to reflect code changes

* Add ParallelHash test vectors and fix problems found.

It turns out that during the initial implementation of ParallelHash an
important detail was missed where if the function name and customization
strings are both the empty string "" then CSHAKE128(X,L,"","") is
equivalent to SHAKE128(X,L). Since ParallelHash internally uses CSHAKE
in such a manner we must instead use Generic_Parallel_XOF instead of
Generic_Parallel_CSHAKE.

* Remove unused Left_Encode_Long_Long function

* Add precondition to cSHAKE to prevent empty strings.

In CSHAKE at least one of the Function_Name or Customization strings
must be non-empty strings. If both strings are empty then cSHAKE is
equivalent to SHAKE, and this case is not handled in this implementation
(since the suffix bits are different between cSHAKE and SHAKE).

* Add ParallelHash to readme

* Update build instructions

* Use --all by default when running the benchmark