Add support for AVX2.
This commit adds support for the AVX2 instruction set for 4x parallelism
with Keccak. The 8x implementation falls back to two 4x instances.

This is currently untested, due to lack of available hardware with which
to test. However, since the same generic code is re-used from the SSE2
implementation and instantiated for the new AVX2 definitions, I'm
confident it should work.