Separate Keccak instantiations in different packages for different round counts.
This change reorganizes the instantiations of the various Keccak
constructs (Keccak-f, Sponge, Duplex) into separate child packages
for different numbers of Keccak-f rounds.
For example, Keccak.Keccak_1600.Rounds_24 contains instantiations
for things based on Keccak-p[1600,24]. That is, Keccak-f[1600]
using 24 rounds. Similarly, Keccak.Keccak_1600.Rounds_12 has
instantiations for Keccak-p and Sponges based on 12 rounds of
Keccak-f[1600].

The purpose of this is to provide cleaner organization of the code.