Add distinct type for lane complemented Keccak-f implementation.
The lane complemented implementation of the Keccak-f permutation
must only be used with specific implementations of subprograms (such as
XOR_Bits_Into_State and Extract_Bytes) in order to obtain correct
results, as those subprograms correctly handle the lane complementing
the required lanes when extracting bits from the Keccak-f state.

This commit adds the distinct type Lane_Complemented_State separate
from the Generic_KeccakF.State type so that using the wrong
subprograms with the lane complemented state will yield a compiler
error, thus ensuring that it can only be used in the correct way.