Replace some single-letter names with more descriptive names. (#11)
In several instances the terminology used is chosen to match the
Keccak specification. E.g. 'F' for the permutation function,
'L' for the lane size logarithm, 'W' for the lane size in bits,
and 'B' for the Keccak state size in bits. However, these names
are not particularly descriptive when reading the source code
unless the reader is familiar with the Keccak specification.

This commit renames these instances to use more descriptive names,
e.g. 'Permute' instead of 'F', as well as clarifying a few other names.