Initial implementation of Ketje v2 (untested).
This commit implements the MonkeyWrap construction on top of
MonkeyDuplex, with four concrete instantiations for Ketje Jr, Sr,
Minor, and Major.

MonkeyWrap is implemented as a streaming API, so each subprogram
can be called multiple times to process a large amount of data
in chunks of arbitrary length.

Tests & benchmarks are still todo.