Fix Gimli bugs.
There was an incorrect assignment of the round constant to the
Gimli state (should be a XOR). There was also an incorrect calculation
of the column number when extracting the final bytes from the state.

Furthermore, there was a missing suffix of four '1' bits in the
definition of Gimli Hash.