Fix endianness of the Ascon padding.
Unlike the existing Pad10*1 rule, Ascon's padding uses a big-endian
convention so the first padding '1' bit is in the MSB of the data.
I.e. the padding is 16#80#, not 16#01#.

This also requires special handling for bit-oriented messages
(which is already supported by Keccak.Generic_Sponge), where the
last partial byte must be shifted to align with the MSB.

Unfortunately there are no Ascon test vectors or reference
implementation for bit-oriented messages to verify that this
behaviour is correct.