Implement Ascon-Hash and Ascon-XOF (#15)
* Implement the Ascon permutation

* Allow some initial data to be passed to Generic_Sponge.Init.

Ascon-Hash requires the ability to initialise the Ascon state with
some specific data. This commit updates Generic_Sponge.Init to be
provided with an additional (optional) parameter to specify this
initial data.

* Implement Ascon-Hash.

This algorithm is very similar to Keccak.Generic_Hash, except
with some minor differences required by the Ascon specification.

* Fix Ascon endianness

* Add Ascon-Hash test vectors

* Implement the Pad*01 rule.

This is similar to the Pad1*01 rule except the first '1' bit is omitted.

* Use Pad*01 rule for Ascon-Hash

* Remove Depends aspect on Keccak.Generic_Sponge.Init.

This seems to generate an erroneous self-dependency on Ctx
(i.e. it expectes that Ctx => Ctx is a dependency), even though Ctx is
an 'out' parameter. As a workaround, the Depends contract is removed
for this subprogram.

* Enable all Ascon test vectors.

Currently, only the test vector with empty input is passing.
There's probably a bug with abosrbing or extracting data.

* Fix incorrect conversion from BE bytes-to-Unsigned_64

* Fix endianness of the Ascon padding.

Unlike the existing Pad10*1 rule, Ascon's padding uses a big-endian
convention so the first padding '1' bit is in the MSB of the data.
I.e. the padding is 16#80#, not 16#01#.

This also requires special handling for bit-oriented messages
(which is already supported by Keccak.Generic_Sponge), where the
last partial byte must be shifted to align with the MSB.

Unfortunately there are no Ascon test vectors or reference
implementation for bit-oriented messages to verify that this
behaviour is correct.

* Instantiate Ascon.Hash from Keccak.Generic_Hash.

This adds an additional formal generic parameter to Keccak.Generic_Hash
to allow the initial state of the permutation to be set. This defaults
to an empty buffer to remain compatible with existing hash functions
which are also instantiated from Keccak.Generic_Hash.

* Instantiate Ascon-XOF.

This makes the same modifications to Keccak.Generic_XOF as were
required to Keccak.Generic_Hash to allow the initial permutation
state to be set in the sponge.

* Update Ascon KAT tests and add Ascon-XOF test vectors.

Now that Ascon-Hash is based on Keccak.Generic_Hash we can
delete the specific AsconHash_Runner and instantiate the
existing Hash_Runner with Ascon.Hash. This required the
test vector format to be modified to fit the schema required
by Hash_Runner.

* Add test vectors for Ascon-XOF.

* Add Ascon to the benchmark

* Fix proof failure

* Document the source of the Ascon test vectors