Update to Alire and GNAT 11 (#18)
This performs several major updates to bring the library up-to-date with the latest GNAT and GNATprove tools, and using the Alire package manager. The summary of changes are:

 * Added `alire.toml`.
 * Regenerated `libkeccak.gpr` with Alire.
 * Refactored the code to resolve errors found with more recent versions of GNAT.
 * Refactored the code to ensure clean proof with the most recent version of GNATprove (11.2.0-3).
 * Regenerated the test projects (benchmark, kat, unit_tests) with Alire.
 * Removed the Makefiles since they are no longer necessary with Alire.
 * Resolved various style issues in the test programs found by style checks.
 * Removed Travis CI. It is planned to replace this with GitHub Actions in a separate pull request.
 * Updated documentation (READMEs).

The library now requires at least GNAT 11 to build due to the use of `Relaxed_Initialization`. An attempt was made to replace all manual suppressions of complex initialization checks (`pragma Annotate`) with `Relaxed_Initialization`, but it has been more difficult than anticipated to get the proofs working due to the complexity of some of the proofs (parallel algorithms in particular). It is therefore planned to tackle this as a separate piece pull request.