Fix incorrect precondition at Keccak.Generic_MonkeyWrap.Init.
The precondition was overly restrictive and was being violated when
running the test vectors for Ketje. The requirement according to
the Ketj2 v2 spec is that the |K| < b - 18 and |N| < b - |K| - 18.
The precondition is updated to reflect the actual requirement.

The changes have been proven with GNATprove, and the KAT suite and
unit tests have been re-run and passed.