Fix incorrect definition for Rotate_Left in SSE2 instantiations.
The definition for `Rotate_Left` in an instantiation of Generic_Parallel_KeccakF
with SSE2 instructions was incorrectly set to `Shift_Left` instead of
`Rotate_Left`, which resulted in KAT test failures which was caught by CI
during the KAT suite.

The unit tests have been expanded slightly to repeat a test over
different data lengths to help catch regressions for this type of error.