Only run proofs in CI when triggered manually
The proofs take a long time to run (about 1h 20m) on the dual-core
CI runners due the large number of verification conditions.
They are therefore changed so that they are only run manually, or
when a release is created.

The purpose of the manual proofs is so that we can choose how the
proofs are run when reviewing pull requests. We can either run
the proofs in CI (which takes a long time), or we can pull the
branch and run the proofs locally on a much more powerful machine
to get the results faster (about 12 minutes per run on my machine).