Move gnatprove dependency from the top-level alire.toml to a nested crate.
This prevents users of this crate from needing to pull in gnatprove
(which is quite large) when they "with" this crate. gnatprove is
used only for verifying the library and is not needed to use it.

The recommended workflow is to have a nested crate which depends
on the library and GNATprove, and that is used for verification.
Having an actual project file does not seem to be needed, since
`alr exec` can be used to run `gnatprove` on `libkeccak.gpr` within
in the nested crate's environment.